Vitis-970 Move mature APIs to include/xrt folder (#5003)
* Move mature APIs to include/xrt folder

* Add wrappers for moved headers for backwards compatibility

* Update include path for moved headers in other exported headers

* Update path to moved headers in doc

* Update public test include paths for moved headers

* Add missing dependency on target that generates rpc_messages.pb.h
