+++ src/runtime_src/core/include/experimental/xrt_aie.h
@@ -20,9 +20,9 @@
 #define _XRT_AIE_H_
 
 #include "xrt.h"
-#include "experimental/xrt_uuid.h"
-#include "experimental/xrt_bo.h"
-#include "experimental/xrt_device.h"
+#include "xrt/xrt_uuid.h"
+#include "xrt/xrt_bo.h"
+#include "xrt/xrt_device.h"
 #include "experimental/xrt_graph.h"
 
 #ifdef __cplusplus
