+++ src/runtime_src/core/include/experimental/xrt_uuid.h
@@ -1,5 +1,6 @@
-/**
- * Copyright (C) 2016-2020 Xilinx, Inc
+/*
+ * Copyright (C) 2021, Xilinx Inc - All rights reserved
+ * Xilinx Runtime (XRT) Experimental APIs
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may
  * not use this file except in compliance with the License. A copy of the
@@ -13,109 +14,4 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-
-#ifndef xrt_uuid_h_
-#define xrt_uuid_h_
-
-#ifdef _WIN32
-# include "windows/uuid.h"
-#else
-# include <uuid/uuid.h>
-typedef uuid_t xuid_t;
-#endif
-
-#ifdef __cplusplus
-#include <string>
-
-namespace xrt {
-
-/**
- * uuid wrapper to treat uuid_t as a value type
- * supports copying
- */
-class uuid
-{
-  xuid_t m_uuid;
-public:
-  uuid()
-  {
-    uuid_clear(m_uuid);
-  }
-
-  uuid(const xuid_t val)
-  {
-    uuid_copy(m_uuid,val);
-  }
-
-  explicit uuid(const std::string& uuid_str)
-  {
-    uuid_parse(uuid_str.c_str(), m_uuid);
-  }
-
-  uuid(const uuid& rhs)
-  {
-    uuid_copy(m_uuid,rhs.m_uuid);
-  }
-
-  uuid(uuid&&) = default;
-  uuid& operator=(uuid&&) = default;
-
-  uuid& operator=(const uuid& rhs)
-  {
-    uuid source(rhs);
-    std::swap(*this,source);
-    return *this;
-  }
-
-  const xuid_t& get() const
-  {
-    return m_uuid;
-  }
-
-  std::string
-  to_string() const
-  {
-    char str[40] = {0};
-    uuid_unparse_lower(m_uuid,str);
-    return str;
-  }
-
-  operator bool() const
-  {
-    return uuid_is_null(m_uuid) == false;
-  }
-
-  bool
-  operator == (const xuid_t& xuid) const
-  {
-    return uuid_compare(m_uuid, xuid) == 0;
-  }
-
-  bool
-  operator != (const xuid_t& xuid) const
-  {
-    return uuid_compare(m_uuid, xuid) != 0;
-  }
-
-  bool
-  operator == (const uuid& rhs) const
-  {
-    return uuid_compare(m_uuid, rhs.m_uuid) == 0;
-  }
-
-  bool
-  operator != (const uuid& rhs) const
-  {
-    return uuid_compare(m_uuid, rhs.m_uuid) != 0;
-  }
-
-  bool
-  operator < (const uuid& rhs) const
-  {
-    return uuid_compare(m_uuid, rhs.m_uuid) < 0;
-  }
-};
-
-} // xrt
-#endif
-#endif
+#include "xrt/xrt_uuid.h"
