+++ src/runtime_src/core/pcie/emulation/cpu_em/CMakeLists.txt
@@ -21,6 +21,7 @@ file(GLOB EM_SRC_FILES
 
 add_definitions(-DXCLHAL_MAJOR_VER=1 -DXCLHAL_MINOR_VER=0)
 add_library(sw_emu_objects OBJECT ${EM_SRC_FILES})
+add_dependencies(sw_emu_objects pcie_emulation_generated_code)
 
 set(CURR_SOURCE "")
 add_library(xrt_swemu SHARED ${CURR_SOURCE}
