+++ tests/unit_test/cuselect/xrtxx.cpp
@@ -17,9 +17,9 @@
 // % g++ -g -std=c++14 -I$XILINX_XRT/include -L$XILINX_XRT/lib -o xrtxx.exe xrtxx.cpp -lxrt_coreutil
 
 
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
-#include "experimental/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
+#include "xrt/xrt_device.h"
 
 #include <fstream>
 #include <stdexcept>
