+++ tests/xrt/00_hello/main.cpp
@@ -20,9 +20,9 @@
 #include <vector>
 #include <cstdlib>
 
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 #include "experimental/xrt_ini.h"
 /**
  * Trivial loopback example which runs OpenCL loopback kernel. Does not use OpenCL
