+++ tests/xrt/04_swizzle/main.cpp
@@ -20,9 +20,9 @@
 #include <string>
 #include <cstring>
 
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 
 static const int DATA_SIZE = 4096;
