+++ tests/xrt/07_sequence/main.cpp
@@ -19,9 +19,9 @@
 #include <string>
 #include <cstring>
 
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 
 /**
  * Runs an OpenCL kernel which writes known 16 integers into a 64 byte
