+++ tests/xrt/102_multiproc_verify/main.cpp
@@ -29,9 +29,9 @@
 #include <unistd.h>
 
 
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 
 namespace {
