+++ tests/xrt/11_fp_mmult256/main.cpp
@@ -21,9 +21,9 @@
 #include <ctime>
 #include <chrono>
 
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 
 #ifdef _WIN32
 # pragma warning ( disable : 4244 )
