+++ tests/xrt/13_add_one/main.cpp
@@ -20,9 +20,9 @@
 #include <numeric>
 #include <cstring>
 
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 
 #define ARRAY_SIZE 8
