+++ tests/xrt/22_verify/main.cpp
@@ -18,9 +18,9 @@
 #include <stdexcept>
 #include <string>
 
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 
 /**
  * Runs an OpenCL kernel which writes "Hello World\n" into the buffer passed
