+++ tests/xrt/56_xclbin/main.cpp
@@ -21,9 +21,9 @@
 
 // XRT includes
 #include "experimental/xrt_xclbin.h"
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 
 // This value is shared with worgroup size in kernel.cl
 constexpr auto COUNT = 1024;
