+++ tests/xrt/fa_kernel/xrt2.cpp
@@ -23,9 +23,9 @@
 #include <chrono>
 #include <numeric>
 
-#include "experimental/xrt_device.h"
-#include "experimental/xrt_kernel.h"
-#include "experimental/xrt_bo.h"
+#include "xrt/xrt_device.h"
+#include "xrt/xrt_kernel.h"
+#include "xrt/xrt_bo.h"
 
 #ifdef _WIN32
 # pragma warning ( disable : 4244 )
