+++ tests/xrt/soft_kernel_hello2/hello_world.cpp
@@ -14,10 +14,10 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-//#include "experimental/xrt_bo.h"
-#include "experimental/xrt_kernel.h"
+//#include "xrt/xrt_bo.h"
+#include "xrt/xrt_kernel.h"
 #include "experimental/xrt_xclbin.h"
-#include "experimental/xrt_device.h"
+#include "xrt/xrt_device.h"
 #include "xrt.h"
 #include <iostream>
 #include <cstring>
